+++ ReSharper.FSharp/test/data/features/intentions/toModuleNamespace/Module 01.fs
@@ -1,0 +1,4 @@
+module Module{caret}
+
+type T() =
+    class end