+++ ReSharper.FSharp/test/data/features/intentions/toModuleNamespace/Namespace 01.fs
@@ -1,0 +1,4 @@
+namespace Ns{caret}
+
+type T() =
+    class end