+++ ReSharper.FSharp/test/data/features/intentions/toModuleNamespace/availability/Module 01.fs
@@ -1,0 +1,1 @@
+module{on} rec Module{on}