+++ ReSharper.FSharp/test/data/features/intentions/toModuleNamespace/availability/Module 02 - Nested.fs
@@ -1,0 +1,4 @@
+namespace Ns
+
+module{off} M{off} =
+  ()