+++ ReSharper.FSharp/test/data/features/intentions/toModuleNamespace/availability/Namespace 01.fs
@@ -1,0 +1,3 @@
+namespace{on} Ns1{on}
+
+namespace{off} global{off}