+++ ReSharper.FSharp/test/src/FSharp.Tests/Intentions/ToModuleNamespaceActionTest.fs
@@ -1,0 +1,22 @@
+namespace JetBrains.ReSharper.Plugins.FSharp.Tests.Features
+
+open JetBrains.ReSharper.Plugins.FSharp.Psi.Features.Intentions
+open NUnit.Framework
+
+type ToModuleNamespaceActionExecuteTest() =
+    inherit FSharpContextActionExecuteTestBase<ToModuleNamespaceDeclarationAction>()
+
+    override x.ExtraPath = "toModuleNamespace"
+
+    [<Test>] member x.``Module 01``() = x.DoNamedTest()
+    [<Test>] member x.``Namespace 01``() = x.DoNamedTest()
+
+
+type ToModuleNamespaceActionAvailabilityTest() =
+    inherit FSharpContextActionAvailabilityTestBase<ToModuleNamespaceDeclarationAction>()
+
+    override x.ExtraPath = "toModuleNamespace"
+
+    [<Test>] member x.``Module 01``() = x.DoNamedTest()
+    [<Test>] member x.``Module 02 - Nested``() = x.DoNamedTest()
+    [<Test>] member x.``Namespace 01``() = x.DoNamedTest()